Refactor stack allocator to use Collection as backend
This eliminates the need for host allocator, storage class, etc. I
refactored the demo interactor detector stuff as part of it, as well as
some of the runner code that needed to be changed as a cosequence.

Additional changes were required in the livermore interactor since
the stack allocator now has a template parameter on memory space.
I refactored some parts of the livermore atomic relaxation to reduce
propagation of those template paramters and to improve program flow.