Use primitive types for IO data
We want to make sure the IO data has consistent types even if we switch
real_type to float. Also, mat_id and its ilk are posing as classes but
are just primitive types, and they have identifiers that conflict with
the use of `mat_id` as variables of type MaterialId elsewhere. Just use
unsigned int to make it clear they're indices.